+++ src/services/voting/SuperElection.ts
@@ -413,7 +413,20 @@ class SuperElection {
   };
 
   combinedApproval(candidates = this.candidates): ResultDetailed {
-    const combinedApprovalResults = {};
+    const initialShape = candidates.reduce((a, c) => ({ ...a, [c]: { score: 0, positive: 0, negative: 0 } }), {});
+    const combinedApprovalResults = Object.values(this.ballotsScored).reduce((a, { ballot, weight }) => {
+      for (let candidate of candidates) {
+        if (ballot[candidate] > (2/3)) {
+          a[candidate].score += weight;
+          a[candidate].positive += weight;
+        } else if (ballot[candidate] < (1/3)) {
+          a[candidate].score -= weight;
+          a[candidate].negative -= weight;
+        }
+      }
+
+      return a;
+    }, initialShape as ResultDetailed);
 
     return combinedApprovalResults;
   };